Create README.md
The NLP Disaster Tweets project involves utilizing Natural Language Processing (NLP) techniques to analyze and classify tweets related to disasters or emergencies. The goal of this project is to develop a model that can accurately identify whether a given tweet is referring to a real disaster or is non-disaster related, often referred to as a "fake" or "not real" disaster tweet.

The project typically involves a dataset consisting of thousands of tweets, where each tweet is labeled as either a disaster or non-disaster tweet. Using NLP techniques, the project aims to extract valuable information and patterns from the text data to classify the tweets correctly.